Minor stylistic suggestions
I took the liberty of directly editing the text where I feel it could be improved. Please feel free to accept/modify/reject whichever change you consider suitable.

My main purpose with these small modifications was to improve clarity just a bit by shortening long sentences and rephrasing a few parts that I considered somewhat strange, but again, it's mostly a stylistic matter and the authors have the last word on this.

I added a `<NAME?>` on line 40 (mid-paragraph) to indicate that, in my view, it looks better if the citation has a proper name, e.g., "one open-source project is Something [1]" instead of just "... is [1]". My rule of thumb is that removing all citations from the text should not make it unreadable or grammatically incorrect. Feel free to decide if this is something you want to fix.

Hope it helps! :smile: